# docs/swagger.yaml
basePath: /
definitions:
  dangbamgong-backend_internal_domain.ErrorCode:
    enum:
    - BAD_REQUEST
    - NOT_FOUND
    - INTERNAL_SERVER_ERROR
    - UNAUTHORIZED
    - FORBIDDEN
    - CONFLICT
    - SERVICE_UNAVAILABLE
    - INVALID_TOKEN
    - INVALID_NICKNAME
    - NICKNAME_ALREADY_SET
    - USER_NOT_FOUND
    - ALREADY_BLOCKED
    - NOT_BLOCKED
    - ALREADY_IN_VOID
    - NOT_IN_VOID
    - TOO_MANY_ACTIVITIES
    - INVALID_ACTIVITY_NAME
    - ACTIVITY_ALREADY_EXISTS
    - ACTIVITY_NOT_FOUND
    - NOTIFICATION_NOT_FOUND
    - ALREADY_FRIENDS
    - REQUEST_ALREADY_SENT
    - BLOCKED
    - REQUEST_NOT_FOUND
    - REQUEST_NOT_PENDING
    - NOT_FRIENDS
    - FRIEND_NOT_IN_VOID
    - INVALID_REQUEST_TYPE
    type: string
    x-enum-varnames:
    - ErrBadRequest
    - ErrNotFound
    - ErrInternalServer
    - ErrUnauthorized
    - ErrForbidden
    - ErrConflict
    - ErrServiceUnavailable
    - ErrInvalidToken
    - ErrInvalidNickname
    - ErrNicknameAlreadySet
    - ErrUserNotFound
    - ErrAlreadyBlocked
    - ErrNotBlocked
    - ErrAlreadyInVoid
    - ErrNotInVoid
    - ErrTooManyActivities
    - ErrInvalidActivityName
    - ErrActivityAlreadyExists
    - ErrActivityNotFound
    - ErrNotificationNotFound
    - ErrAlreadyFriends
    - ErrRequestAlreadySent
    - ErrBlocked
    - ErrRequestNotFound
    - ErrRequestNotPending
    - ErrNotFriends
    - ErrFriendNotInVoid
    - ErrInvalidRequestType
...
  dangbamgong-backend_internal_dto.BucketItem:
    properties:
      count:
        type: integer
      isMine:
        type: boolean
      time:
        type: string
    type: object
...
  dangbamgong-backend_internal_dto.DailyStatResponse:
    properties:
      buckets:
        items:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.BucketItem'
        type: array
      mySessions:
        items:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.VoidSessionItem'
        type: array
      targetDay:
        type: string
    type: object
  dangbamgong-backend_internal_dto.ErrorResponse:
    properties:
      code:
        $ref: '#/definitions/dangbamgong-backend_internal_domain.ErrorCode'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.FriendItem:
    properties:
      createdAt:
        type: string
      isInVoid:
        type: boolean
      lastVoidEndedAt:
        type: string
      nickname:
        type: string
      tag:
        type: string
      userId:
        type: string
    type: object
  dangbamgong-backend_internal_dto.FriendListResponse:
    properties:
      friends:
        items:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.FriendItem'
        type: array
    type: object
  dangbamgong-backend_internal_dto.HomeStatResponse:
    properties:
      currentVoidCount:
        description: 공통
        type: integer
      myRank:
        type: integer
      myTotalDurationSec:
        description: 오늘 잠에 든 유저
        type: integer
      todaySleptCount:
        type: integer
      totalSleptUsers:
        type: integer
    type: object
...
  dangbamgong-backend_internal_dto.MyVoidStatResponse:
    properties:
      averageDurationSec:
        type: integer
      maxDurationSec:
        type: integer
      totalDurationSec:
        type: integer
    type: object
  dangbamgong-backend_internal_dto.NotificationItem:
    properties:
      body:
        type: string
      createdAt:
        type: string
      id:
        type: string
      isRead:
        type: boolean
      title:
        type: string
      type:
        type: string
    type: object
  dangbamgong-backend_internal_dto.NotificationListResponse:
    properties:
      hasMore:
        type: boolean
      notifications:
        items:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.NotificationItem'
        type: array
    type: object
...
  dangbamgong-backend_internal_dto.ReceivedRequestItem:
    properties:
      createdAt:
        type: string
      requestId:
        type: string
      sender:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.UserSearchItem'
    type: object
  dangbamgong-backend_internal_dto.ReceivedRequestsResponse:
    properties:
      requests:
        items:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.ReceivedRequestItem'
        type: array
    type: object
...
  dangbamgong-backend_internal_dto.Response-any:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_DailyStatResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.DailyStatResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_FriendListResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.FriendListResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_HomeStatResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.HomeStatResponse'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_MyVoidStatResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.MyVoidStatResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_NotificationListResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.NotificationListResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_ReceivedRequestsResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.ReceivedRequestsResponse'
      success:
        type: boolean
    type: object
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_SendFriendRequestResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.SendFriendRequestResponse'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_UnreadCountResponse:
    properties:
      data:
        $ref: '#/definitions/dangbamgong-backend_internal_dto.UnreadCountResponse'
      success:
        type: boolean
    type: object
...
  dangbamgong-backend_internal_dto.SendFriendRequestRequest:
    properties:
      receiverId:
        type: string
    required:
    - receiverId
    type: object
  dangbamgong-backend_internal_dto.SendFriendRequestResponse:
    properties:
      requestId:
        type: string
    type: object
...
  dangbamgong-backend_internal_dto.UnreadCountResponse:
    properties:
      count:
        type: integer
    type: object
...
  dangbamgong-backend_internal_dto.UserSearchItem:
    properties:
      isBlocked:
        type: boolean
      nickname:
        type: string
      tag:
        type: string
      userId:
        type: string
    type: object
...
  dangbamgong-backend_internal_dto.VoidSessionItem:
    properties:
      activities:
        items:
          type: string
        type: array
      endedAt:
        type: string
      startedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 밤의 공백 시간 측정 소셜 앱 API
  title: 당밤공 API
  version: "1.0"
paths:
...
  /friends:
    get:
      description: 친구 목록을 반환합니다. 각 친구의 공백 상태를 포함합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_FriendListResponse'
      security:
      - BearerAuth: []
      summary: 친구 목록 조회
      tags:
      - Friends
  /friends/{user_id}:
    delete:
      description: 양방향 친구 관계를 삭제합니다
      parameters:
      - description: 삭제할 친구 유저 ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "404":
          description: NOT_FRIENDS
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 친구 삭제
      tags:
      - Friends
  /friends/{user_id}/nudge:
    post:
      description: 공백 중인 친구에게 알림을 보냅니다. 친구가 공백 중이 아니면 실패합니다.
      parameters:
      - description: 찌를 친구 유저 ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "400":
          description: NOT_FRIENDS / FRIEND_NOT_IN_VOID
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 친구 찌르기
      tags:
      - Friends
  /friends/requests:
    get:
      description: 받은(received) 또는 보낸(sent) 친구 요청 목록을 반환합니다
      parameters:
      - description: 요청 타입
        enum:
        - received
        - sent
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: type=received
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_ReceivedRequestsResponse'
        "400":
          description: INVALID_REQUEST_TYPE
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 친구 요청 목록 조회
      tags:
      - Friends
    post:
      consumes:
      - application/json
      description: 유저에게 친구 요청을 보냅니다. 이미 친구이거나 차단 관계이면 실패합니다.
      parameters:
      - description: 받는 유저 ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dangbamgong-backend_internal_dto.SendFriendRequestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_SendFriendRequestResponse'
        "400":
          description: BLOCKED
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
        "409":
          description: ALREADY_FRIENDS / REQUEST_ALREADY_SENT
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 친구 요청 보내기
      tags:
      - Friends
  /friends/requests/{request_id}:
    delete:
      description: 보낸 친구 요청을 삭제합니다. 이미 수락된 요청은 삭제할 수 없습니다.
      parameters:
      - description: 요청 ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "404":
          description: REQUEST_NOT_FOUND
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 친구 요청 삭제
      tags:
      - Friends
  /friends/requests/{request_id}/accept:
    post:
      description: 받은 친구 요청을 수락합니다. 양방향 친구 관계가 생성됩니다.
      parameters:
      - description: 요청 ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "400":
          description: REQUEST_NOT_PENDING
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
        "404":
          description: REQUEST_NOT_FOUND
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 친구 요청 수락
      tags:
      - Friends
  /friends/requests/{request_id}/reject:
    post:
      description: 받은 친구 요청을 거절합니다. 요청은 삭제되지 않고 상태만 변경됩니다.
      parameters:
      - description: 요청 ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "400":
          description: REQUEST_NOT_PENDING
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
        "404":
          description: REQUEST_NOT_FOUND
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 친구 요청 거절
      tags:
      - Friends
...
  /notifications:
    get:
      description: 유저의 알림 목록을 페이지네이션으로 반환합니다
      parameters:
      - description: 조회 개수 (기본 20)
        in: query
        name: limit
        type: integer
      - description: 오프셋
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_NotificationListResponse'
      security:
      - BearerAuth: []
      summary: 알림 목록 조회
      tags:
      - Notifications
  /notifications/{notification_id}/read:
    patch:
      description: 특정 알림을 읽음 상태로 변경합니다
      parameters:
      - description: 알림 ID
        in: path
        name: notification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-any'
        "404":
          description: NOTIFICATION_NOT_FOUND
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 알림 읽음 처리
      tags:
      - Notifications
  /notifications/unread-count:
    get:
      description: 읽지 않은 알림의 개수를 반환합니다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_UnreadCountResponse'
      security:
      - BearerAuth: []
      summary: 읽지 않은 알림 수 조회
      tags:
      - Notifications
  /stats/daily:
    get:
      description: 특정 날짜의 20분 단위 버킷 통계와 내 공백 세션 목록을 반환합니다
      parameters:
      - description: 조회할 날짜 (YYYY-MM-DD)
        in: query
        name: target_day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_DailyStatResponse'
      security:
      - BearerAuth: []
      summary: 일별 통계 조회
      tags:
      - Stats
  /stats/home:
    get:
      description: 현재 공백 중인 유저 수, 오늘 잠에 든 유저 수, 내 랭킹/총 공백 시간을 반환합니다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_HomeStatResponse'
      security:
      - BearerAuth: []
      summary: 홈 통계 조회
      tags:
      - Stats
  /stats/me:
    get:
      description: 전체 공백 총 시간, 평균 시간, 최대 시간을 반환합니다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dangbamgong-backend_internal_dto.Response-dangbamgong-backend_internal_dto_MyVoidStatResponse'
      security:
      - BearerAuth: []
      summary: 내 공백 통계 조회
      tags:
      - Stats
...
securityDefinitions:
  BearerAuth:
    description: Bearer {JWT 토큰}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"